FIXED: import, types, and removed query params from components
