+++ src/pages/hes/live-data/block-load/index.tsx
@@ -1,24 +1,10 @@
-import { useSearchParams } from 'react-router-dom';
-import BlockLoadTable from './includes/blockLoadTable';
+import BlockLoadTable from './includes/BlockLoadTable';
 import { useState } from 'react';
-import BlockLoadGraph from './includes/blockloadGraph';
+import BlockLoadGraph from './includes/BlockloadGraph';
 import ToggleView from '@/components/customUI/ToggleView';
 const BlockLoad = () => {
-  const [searchParams, setSearchParams] = useSearchParams();
+  
   const [view, setView] = useState<string>('graph');
-  const mapParams = (params: URLSearchParams) => {
-    const mappedParams = new URLSearchParams();
-    if (params.has('site_ids'))
-      mappedParams.set('site_id', params.get('site_ids')!);
-    if (params.has('pss_ids'))
-      mappedParams.set('pss_id', params.get('pss_ids')!);
-    if (params.has('feeder_ids'))
-      mappedParams.set('feeder_id', params.get('feeder_ids')!);
-    if (params.has('dtr_ids'))
-      mappedParams.set('dtr_id', params.get('dtr_ids')!);
-    return mappedParams;
-  };
-  const mappedSearchParams = mapParams(searchParams);
 
   return (
     <div className="px-5 py-3 w-full">
@@ -31,10 +17,7 @@ const BlockLoad = () => {
         </div>
         <div className="overflow-x-scroll">
           {view === 'table' && (
-            <BlockLoadTable
-              groupName={'Block Load'}
-              search={mappedSearchParams.toString()}
-            />
+            <BlockLoadTable />
           )}
           {view === 'graph' && <BlockLoadGraph />}
         </div>
