+++ src/pages/hes/live-data/daily-load/includes/DailyLoadGraph.tsx
