+++ src/pages/hes/live-data/daily-load/index.tsx
@@ -1,24 +1,11 @@
-import { useSearchParams } from 'react-router-dom';
 import { useState } from 'react';
-import DailyLoadTable from './includes/dailyLoadTable';
-import DailyLoadGraph from './includes/dailyLoadGraph';
+import DailyLoadTable from './includes/DailyLoadTable';
+import DailyLoadGraph from './includes/DailyLoadGraph';
 import ToggleView from '@/components/customUI/ToggleView';
+
 const DailyLoad = () => {
-  const [searchParams, setSearchParams] = useSearchParams();
+
   const [view, setView] = useState<string>('graph');
-  const mapParams = (params: URLSearchParams) => {
-    const mappedParams = new URLSearchParams();
-    if (params.has('site_ids'))
-      mappedParams.set('site_id', params.get('site_ids')!);
-    if (params.has('pss_ids'))
-      mappedParams.set('pss_id', params.get('pss_ids')!);
-    if (params.has('feeder_ids'))
-      mappedParams.set('feeder_id', params.get('feeder_ids')!);
-    if (params.has('dtr_ids'))
-      mappedParams.set('dtr_id', params.get('dtr_ids')!);
-    return mappedParams;
-  };
-  const mappedSearchParams = mapParams(searchParams);
 
   return (
     <div className="px-5 py-3 w-full">
@@ -31,10 +18,7 @@ const DailyLoad = () => {
         </div>
         <div className="overflow-x-scroll">
           {view === 'table' && (
-            <DailyLoadTable
-              groupName={'Block Load'}
-              search={mappedSearchParams.toString()}
-            />
+            <DailyLoadTable />
           )}
           {view === 'graph' && <DailyLoadGraph />}
         </div>
