+++ src/pages/hes/live-data/instantaneous-profile/index.tsx
@@ -1,25 +1,12 @@
-import { useSearchParams } from 'react-router-dom';
 import { useState } from 'react';
 import InstantaneousGraph from './includes/InstantaneousGraph';
 import ToggleView from '@/components/customUI/ToggleView';
 import InstantaneousTable from './includes/InstantaneousTable';
-const InstantaneousProfile = () => {
-  const [searchParams, setSearchParams] = useSearchParams();
-  const [view, setView] = useState<string>('graph');
-  const mapParams = (params: URLSearchParams) => {
-    const mappedParams = new URLSearchParams();
-    if (params.has('site_ids'))
-      mappedParams.set('site_id', params.get('site_ids')!);
-    if (params.has('pss_ids'))
-      mappedParams.set('pss_id', params.get('pss_ids')!);
-    if (params.has('feeder_ids'))
-      mappedParams.set('feeder_id', params.get('feeder_ids')!);
-    if (params.has('dtr_ids'))
-      mappedParams.set('dtr_id', params.get('dtr_ids')!);
-    return mappedParams;
-  };
-  const mappedSearchParams = mapParams(searchParams);
 
+const InstantaneousProfile = () => {
+
+  const [view, setView] = useState<string>('graph');
+ 
   return (
     <div className="px-5 py-3 w-full">
       <div className="flex relative flex-col mt-8">
@@ -33,10 +20,7 @@ const InstantaneousProfile = () => {
         </div>
         <div className="overflow-x-scroll">
           {view === 'table' && (
-            <InstantaneousTable
-              groupName={'Block Load'}
-              search={mappedSearchParams.toString()}
-            />
+            <InstantaneousTable />
           )}
           {view === 'graph' && <InstantaneousGraph />}
         </div>
