+++ src/store/hes/endpoints/live-data/index.ts
@@ -10,7 +10,6 @@ import { DailyLoadEndPoints } from './dailyload';
 import { MonthlyBillingEndPoints } from './monthlyBilling';
 import { LiveDataMetricsResponse } from '../../types';
 import {
-  LiveDataMetricsRecord,
   ModifiedLiveDataRecord
 } from '../../types/live-data-metrics';
 import { InstantaneousProfileEndpoints } from './instantaneousProfile';
